+++ dialogs.py
@@ -78,6 +78,9 @@ class UjdonsagokDialog(wx.Dialog):
         
         ujdonsagok_lista = [
             "Javítva a listafrissítés hibája a Dezideráta-kezelőben, mely akkor jelentkezett, ha a könyv szerkesztése dialog az adatlapról lett megnyitva.",
+            "Javítva a statisztika kereszttáblás jelentése, mely eddig nem tartalmazta az elsődleges szempont értékét.",
+            "Javítva a frissítés keresésekor felugró hibaüzenet internetkapcsolat hiánya esetén.",
+            "Javítva a téma alkalmazása a beállítások ablak görgetésekor.",
             "Újabb kódrefaktorálás és javítások."
         ]